mvp_pygeoapi_logging_demo/server/requirements.txt
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http
scikit-learn==1.5.2
geopandas==1.0.1
numpy==1.22